chore(core): prevent uncommitted changes when building core package (#276)
When building the core package, as part of the generation step the package.json file is generated with no ending new line.

This causes uncommitted changes to be created for the package.json file due to it being committed with a newline by prettier already. By adding a new line upon generation, we should remove this.

This makes it possible to do continual deployment.